test(settings): add API tests for user settings retrieval and validation
